Fixed --fixationmethod gwF implementation
Previously, the re-scaled prefs were not
normalized by the sum of the re-scaled prefs
before calculating fixation probs. This has
been fixed. It is only import for gwF because
in the other methods (HalpernBruno, FracTolerated),
we always have a ratio of re-scaled prefs and
so normalization by sum drops out.