Accept underflow when computed `M` and `dM`.
In the substitution probs `M` and its derivatives `dM`,
underflow will sometimes lead to substitution probabilities
of 0. Before this caused an error, but now it is accepted.
This is because (particularly in codon model), some
transitions are indeed accepted to have near-zero values
(such as between codons that differ by > 1 nucleotide).