Bug fixes to `ExpCM_empirical_phi_divpressure`
Major bug was that the `_update_Frxy` method only did
anything if `omega2` was a free parameter. But in fact
`omega2` needs to be used in the calculation even if it
is **not** a free parameter as it is still in equations
even if it has a constant value.

Other bug was in documentation, which said that `divPressureValues`
was a `list`. This is incorrect, it needs to be a `scipy.array` since
the `__init__` method does `divPressureValues.copy()` which
is valid for a `scipy.array` but not a `list`.