Relaxed test for negative `M(t)`
The `M(t)` values should always be non-negative. But
sometimes they end up being a bit less than zero. I
*think* this is due to numerical error in the matrix
exponentiation, and it is probably for values that
don't really matter as they are for low probability
mutations. But I am not sure. I have changed the
assertion message to a warning error that prints out
some diagnostic information.