Error if Jacobian too large after optimization
For reasons that remain unclear to me, sometimes
the optimizer fails to take a step even when
there is a very large Jacobian. This problem does
**not** seem fixable by just re-seeding model params
as problem persists, so maybe it has something to
do with pathologically wrong branch lengths.

In the absence of a true fix, this commit at least
raises an error if these is an extremely large
Jacobian after optimization has supposedly converged.