changed bullet point formatting
The `.rst` formatting for bullet points was not showing
up in the "command line usage" portion of the documentation.
I changed the formatting from "* *YNGKP_M0*" to "-*YNGKP_M0"
to force a bullet point notation.

I checked through the auxillary programs and it seems like
this was the only page with this bug.