added another test for random seed setting
Previously, this test looked at whether or not you
get the same alignment when you send the random seed
into the function.

I've added another test which tests whether or not you
get the same alignments if you seed the seed before a for
loop but don't send a seed into the simulate function.