fix versions for Travis tests (no pinning)
The Travis tests for Python 3.6 and 3.7 were failing
because package versions were too old. The main cause
was that the tests pinned exact versions in a
`test_requirements.txt` file. This has now been eliminated,
and the latest package versions are just installed from
the `setup.py` requirements.

In addition, the minimal `cython` version requirement has
been increased as required for python 3.7.